Refactored ObjectExtensions
Split the extensions into separate files and started to rewrite
functions that have multiple exit paths
