+++ letsencrypt-win-simple/Extensions/ArrayExtensions.cs
@@ -1,0 +1,13 @@
+﻿using System;
+namespace LetsEncrypt.ACME.Simple.Extensions
+{
+    public static class ArrayExtensions
+    {
+        public static void ForEach(this Array array, Action<Array, int[]> action)
+        {
+            if (array.LongLength == 0) return;
+            ArrayTraverse walker = new ArrayTraverse(array);
+            do action(array, walker.Position); while (walker.Step());
+        }
+    }
+}
