+++ letsencrypt-win-simple/Extensions/ObjectExtensions.cs
@@ -1,0 +1,95 @@
+﻿using System;
+using System.Linq;
+using System.Collections.Generic;
+using System.Reflection;
+using LetsEncrypt.ACME.Simple.Extensions;
+namespace LetsEncrypt.ACME.Simple.Extensions
+{
+    //From https://github.com/Burtsev-Alexey/net-object-deep-copy
+    public static class ObjectExtensions
+    {
+        private static readonly MethodInfo CloneMethod = typeof(Object).GetMethod("MemberwiseClone",
+            BindingFlags.NonPublic | BindingFlags.Instance);
+
+        public static bool IsPrimitive(this Type type)
+        {
+            return type == typeof(String) || (type.IsValueType & type.IsPrimitive);
+        }
+
+        public static Object Copy(this Object originalObject)
+        {
+            return InternalCopy(originalObject, new Dictionary<Object, Object>(new ReferenceEqualityComparer()));
+        }
+
+        private static Object InternalCopy(Object originalObject, IDictionary<Object, Object> visited)
+        {
+            Object result = null;
+            if (originalObject != null)
+            {
+                var typeToReflect = originalObject.GetType();
+                if (IsPrimitive(typeToReflect))
+                {
+                    result = originalObject;
+                }
+                else
+                {
+                    if (!visited.TryGetValue(originalObject, out result))
+                    {
+                        if (!typeof(Delegate).IsAssignableFrom(typeToReflect))
+                        {
+                            var cloneObject = CloneMethod.Invoke(originalObject, null);
+                            if (typeToReflect.IsArray)
+                            {
+                                var arrayType = typeToReflect.GetElementType();
+                                if (IsPrimitive(arrayType) == false)
+                                {
+                                    Array clonedArray = (Array)cloneObject;
+                                    clonedArray.ForEach(
+                                        (array, indices) =>
+                                            array.SetValue(InternalCopy(clonedArray.GetValue(indices), visited), indices));
+                                }
+                            }
+                            visited.Add(originalObject, cloneObject);
+                            CopyFields(originalObject, visited, cloneObject, typeToReflect);
+                            RecursiveCopyBaseTypePrivateFields(originalObject, visited, cloneObject, typeToReflect);
+                            result = cloneObject;
+                        }
+                    }
+                }
+            }
+            return result;
+        }
+
+        private static void RecursiveCopyBaseTypePrivateFields(object originalObject,
+            IDictionary<object, object> visited, object cloneObject, Type typeToReflect)
+        {
+            if (typeToReflect.BaseType != null)
+            {
+                RecursiveCopyBaseTypePrivateFields(originalObject, visited, cloneObject, typeToReflect.BaseType);
+                CopyFields(originalObject, visited, cloneObject, typeToReflect.BaseType,
+                    BindingFlags.Instance | BindingFlags.NonPublic, info => info.IsPrivate);
+            }
+        }
+
+        private static void CopyFields(object originalObject, IDictionary<object, object> visited, object cloneObject,
+            Type typeToReflect,
+            BindingFlags bindingFlags =
+                BindingFlags.Instance | BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.FlattenHierarchy,
+            Func<FieldInfo, bool> filter = null)
+        {
+            foreach (FieldInfo fieldInfo in typeToReflect.GetFields(bindingFlags))
+            {
+                if (filter != null && filter(fieldInfo) == false) continue;
+                if (IsPrimitive(fieldInfo.FieldType)) continue;
+                var originalFieldValue = fieldInfo.GetValue(originalObject);
+                var clonedFieldValue = InternalCopy(originalFieldValue, visited);
+                fieldInfo.SetValue(cloneObject, clonedFieldValue);
+            }
+        }
+
+        public static T Copy<T>(this T original)
+        {
+            return (T)Copy((Object)original);
+        }
+    }
+}
