+++ letsencrypt-win-simple/Extensions/ReferenceEqualityComparer.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+namespace LetsEncrypt.ACME.Simple.Extensions
+{
+    public class ReferenceEqualityComparer : EqualityComparer<Object>
+    {
+        public override bool Equals(object x, object y)
+        {
+            return ReferenceEquals(x, y);
+        }
+
+        public override int GetHashCode(object obj)
+        {
+            if (obj == null) return 0;
+            return obj.GetHashCode();
+        }
+    }
+}
