+++ letsencrypt-win-simple/Plugin/IISSiteServerPlugin.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Security.Cryptography.X509Certificates;
 using Serilog;
-
+using LetsEncrypt.ACME.Simple.Extensions;
 namespace LetsEncrypt.ACME.Simple
 {
     public class IISSiteServerPlugin : Plugin
